feat: add PostViewModel for thread posting
